+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - Purchase order : remove M2O invoice field.
 - STOCKMOVE : display qty per tracking number and not total available qty for tracking number
 - HR : Fix typo.
+- INVOICING PROJECT : Filter the records including deadlineDate.
 
 ## [5.2.1] - 2019-12-16
 ## Features